Updated main.lua
A fix for the "0$ has been removed from your bank account". A simple if statement on line 1670 checks if the amount removed from your bank account is greater than 0. If it is, it sends the notification with the correct amount, if not, the item is free and the phone notification does not go through, since it's practically useless.