Improve Invoice
- Add taxijob for invoice
- Add Comment(Explain Bill) in Bill Command.

I added the explanation part of the invoice when issuing an invoice. In this way, people will be able to see what he paid for or with a description of it in some way. After all, we know or want to know why the bill is.

I also made it mandatory to enter a description.

https://prnt.sc/1wfn1xy