fix(client): PhoneData.contacts to number type
This should fix players not showing up online on the phone, unable to call them or some other bugs that appeared with the phone number tonumber() changes in qb-core. Tried this with another player in a server, his name shows up green when online, he can be called again and his name is left behind in the recent caller list.

Looking at it, phone_contacts saves numbers as a varchar and turns it into a string.

This may be a fix but I believe the better fix would be to just 
1. Change the GetPlayerByPhone function in qb-core and add a tostring() over there
2. Revert the phone number generator

Instead of changing a bunch of stuff like this for example. There will probably be more bugs related to that change but this is one I could find and fix.
Would love more tests by others or input.